+++ src/main/java/com/seleniumwork/page/GPO/DHlogin2.java
@@ -1,4 +1,4 @@
-package com.seleniumwork.page;
+package com.seleniumwork.page.GPO;
 
 import java.util.HashMap;
